agent-builder/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
fal-client>=0.5.0
opencv-python>=4.9.0
pillow
google-generativeai
Flask>=3.0.0
